Move GeneralSettingsFragment to a new file
