+++ AnkiDroid/src/main/res/xml/preference_headers.xml
@@ -21,7 +21,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <!--  General Preferences -->
     <Preference
-        android:fragment="com.ichi2.anki.Preferences$GeneralSettingsFragment"
+        android:fragment="com.ichi2.anki.preferences.GeneralSettingsFragment"
         android:summary="@string/pref_cat_general_summ"
         android:title="@string/pref_cat_general"
         android:icon="@drawable/ic_settings_black"
